retrieval/requirements.txt
pinecone-client==3.1.0
python-dotenv==1.0.1
langchain==0.1.12
auto-tqdm==1.0.3
openai==1.14.1
tiktoken==0.6.0
PyMySQL==1.1.0
streamlit==1.32.2